requirements.txt
apache-beam==2.61.0
apache-flink==2.0.0
apache-flink-libraries==2.0.0
async-timeout==5.0.1
attrs==25.3.0
avro==1.12.0
certifi==2025.6.15
charset-normalizer==3.4.2
cloudpickle==2.2.1
crcmod==1.7
dill==0.3.1.1
dnspython==2.7.0
docopt==0.6.2
fastavro==1.11.1
fasteners==0.19
find_libpython==0.4.1
grpcio==1.65.5
grpcio-tools==1.73.0
hdfs==2.7.3
httplib2==0.22.0
idna==3.10
jsonpickle==3.4.2
jsonschema==4.24.0
jsonschema-specifications==2025.4.1
kafka-python==2.2.12
numpy==2.0.2
objsize==0.7.1
orjson==3.10.18
packaging==25.0
pandas==2.3.0
pemja==0.4.1
proto-plus==1.26.1
protobuf==5.29.5
py4j==0.10.9.7
pyarrow==16.1.0
pyarrow-hotfix==0.7
pydot==1.4.2
PyJWT==2.9.0
pymongo==4.13.2
pyparsing==3.2.3
python-dateutil==2.9.0.post0
pytz==2025.2
PyYAML==6.0.2
redis==5.3.0
referencing==0.36.2
regex==2024.11.6
requests==2.32.4
rpds-py==0.25.1
ruamel.yaml==0.18.14
ruamel.yaml.clib==0.2.12
six==1.17.0
sortedcontainers==2.4.0
typing_extensions==4.14.0
tzdata==2025.2
urllib3==2.5.0
zstandard==0.23.0
